day refactor
changed the package name from day to cooking as it's a slightly better descriptor. I'm still struggling to come up with these one-word package names, though..

also got rid of the gorm dependency as it was simply too difficult to work with. I'm now using the 'database/sql' package, which seems to be much better.